Performed some clean up and restructuring of code
