--- ConsoleApp/IConsoleService.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-
-namespace ConsoleApp
-{
-	public interface IConsoleService : IDisposable
-	{
-		Task RunAsync(string[] args);
-	}
-}
