--- ConsoleApp/IDisposableService.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-
-namespace ConsoleApp
-{
-	public interface IDisposableService : IDisposable
-	{
-		void DoSomething();
-	}
-}
