+++ ConsoleApp/Program.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using ConsoleApp.Services;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using System.Threading.Tasks;
 
@@ -10,18 +11,18 @@ namespace ConsoleApp
 		static async Task Main(string[] args)
 		{
 			IHost host = Host.CreateDefaultBuilder()
-								.ConfigureServices((context, services) =>
-								{
-									services.Configure<CustomOptions>(context.Configuration
-																							.GetSection(CustomOptions.Section));
+				.ConfigureServices((context, services) =>
+				{
+					services.Configure<CustomOptions>(context.Configuration
+																	.GetSection(CustomOptions.Section));
 
-									services.AddTransient<IDisposableService, DisposableService>();
-									services.AddTransient<IConsoleService, ConsoleService>();
+					services.AddTransient<IDisposableService, DisposableService>();
+					services.AddTransient<IConsoleService, ConsoleService>();
 
-									services.AddLogging();
-									services.AddOptions();
-								})
-								.Build();
+					services.AddLogging();
+					services.AddOptions();
+				})
+				.Build();
 
 			await host.Services.GetService<IConsoleService>().RunAsync(args);
