+++ ConsoleApp/Services/ConsoleService.cs
@@ -1,0 +1,42 @@
+﻿using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using System;
+using System.Threading.Tasks;
+
+
+namespace ConsoleApp.Services
+{
+	public class ConsoleService : IConsoleService
+	{
+		private readonly IDisposableService _service;
+		private readonly ILogger<ConsoleService> _logger;
+		private readonly CustomOptions _options;
+
+
+		public ConsoleService(IDisposableService service,
+									ILogger<ConsoleService> logger,
+									IOptions<CustomOptions> options)
+		{
+			_service = service;
+			_logger = logger;
+			_options = options.Value;
+		}
+
+
+		public void Dispose()
+		{
+			_logger.LogInformation("Console.Dispose() has been called.");
+			GC.SuppressFinalize(this);
+		}
+
+
+		public async Task RunAsync(string[] args)
+		{
+			_logger.LogInformation("The application is now starting...");
+			_logger.LogInformation("The Url in the settings file is '{Url}' and use HTTPS is '{UseHttps}'.",
+				_options.Url, _options.UseHttps);
+
+			await _service.DoSomethingAsync();
+		}
+	}
+}
