+++ ConsoleApp/Services/IConsoleService.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Threading.Tasks;
+
+
+namespace ConsoleApp.Services
+{
+	public interface IConsoleService : IDisposable
+	{
+		Task RunAsync(string[] args);
+	}
+}
